+++ src/main/kotlin/main.kt
@@ -1,0 +1,10 @@
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
+
+fun main() {
+    GlobalScope.launch {
+        Client.webClient.run()
+    }
+    consoleUi.LoginMenu().mainAction()
+    consoleUi.MainMenu().mainAction()
+}
